+++ src/main/java/net/silentchaos512/gear/api/material/modifier/IMaterialModifierType.java
@@ -5,8 +5,10 @@ import net.silentchaos512.gear.api.material.IMaterialInstance;
 
 import javax.annotation.Nullable;
 
-public interface IMaterialModifierType {
+public interface IMaterialModifierType<T extends IMaterialModifier> {
     void removeModifier(ItemStack stack);
 
     @Nullable IMaterialModifier read(IMaterialInstance material);
+
+    void write(T modifier, ItemStack stack);
 }